Connect during initialization
There is a time window between `proc_lib:init_ack` is called and
the first ?MSG_CONNECT is handled. Requests sent in that time
window will fail due to no_server error because the counter for
the target server would be 0.

This commit calls the handler of ?MSG_CONNECT in init immidiately
so that when `proc_lib:init_ack` is called, the connections are
already made, or we have already set reconnect timers.